Organise more packages and class, and improve class names
