+++ src/main/java/duke/Duke.java
@@ -1,11 +1,12 @@
 package duke;
 
+import duke.functions.Functions;
+import duke.gui.Window;
+import duke.storage.Storage;
+import duke.storage.TaskList;
 import javafx.application.Application;
 import javafx.scene.Scene;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.layout.HBox;
 import javafx.scene.layout.Pane;
-import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
 
 /**
@@ -22,7 +23,6 @@ public class Duke extends Application {
      * Represents a Duke program.
      */
     private Functions fn;
-    private Scene scene;
 
     public Duke() {
 
@@ -39,52 +39,19 @@ public class Duke extends Application {
             TaskList tl = st.load();
             this.fn = new Functions(tl, st);
         } catch (Exception e) {
-            System.out.println(e);
+            e.printStackTrace();
         }
     }
 
     public Functions getFn() {
-        return fn;
+        return this.fn;
     }
 
     @Override
     public void start(Stage stage) {
-        Duke duke = new Duke("tasks.txt");
-
-        //pane to gather user input
-        Pane inputLayout = new Pane();
-
-        //pane to hold list of tasks
-        VBox outputLayout = new VBox();
-
-        duke.getFn().setOutputLayout(outputLayout);
-
-        //pane to display all task
-        ScrollPane scrollPane = new ScrollPane();
-        scrollPane.setContent(outputLayout);
-        scrollPane.setPrefSize(385, 200);
-        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
-        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
-        scrollPane.setVvalue(1.0);
-        scrollPane.setFitToWidth(true);
-        outputLayout.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
-
-        //pane to hold 3 buttons for each task
-        UI ui = new UI(duke.getFn());
-        HBox buttonLayout = ui.createButtons(inputLayout, outputLayout);
-
-        //main layout
-        VBox mainLayout = new VBox();
-        mainLayout.getChildren().addAll(buttonLayout, inputLayout, scrollPane);
-        mainLayout.setPrefSize(400.0, 600.0);
-        scene = new Scene(mainLayout);
-
-        //set stage
-        stage.setScene(scene);
-        stage.show();
-        stage.setTitle("Duke");
-        stage.setResizable(false);
-        stage.setMinHeight(600.0);
-        stage.setMinWidth(400.0);
+        Duke d = new Duke("tasks.txt");
+        Pane p = Window.makeWindow(d);
+        Scene scene = new Scene(p);
+        Window.setStage(stage, scene);
     }
 }
