+++ src/main/java/duke/dukeExceptions/DeleteException.java
@@ -1,0 +1,10 @@
+package duke.dukeExceptions;
+
+public class DeleteException extends DukeException {
+    /**
+     * Exception used by delete function to indicate invalid user input
+     */
+    public DeleteException() {
+        super("I don't know which task to delete...");
+    }
+}
