+++ src/main/java/duke/dukeExceptions/DukeException.java
@@ -1,0 +1,27 @@
+package duke.dukeExceptions;
+
+public class DukeException extends Exception {
+    /**
+     * General Duke exception. Used for testing.
+     *
+     * @param message Print out any message entered
+     */
+
+    public DukeException(String message) {
+        super(message);
+    }
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
