+++ src/main/java/duke/dukeExceptions/MarkException.java
@@ -1,0 +1,11 @@
+package duke.dukeExceptions;
+
+public class MarkException extends DukeException {
+    /**
+     * Exception used by marking function to indicate that user input is invalid
+     */
+
+    public MarkException() {
+        super("I don't know which task to mark...");
+    }
+}
