+++ src/main/java/duke/storage/TaskList.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.storage;
+
+import duke.task.Task;
 
 import java.util.ArrayList;
 import java.util.List;
