+++ src/test/java/duke/FunctionStub.java
@@ -1,6 +1,10 @@
 package duke;
 
-public class FunctionStub extends Functions{
+import duke.functions.Functions;
+import duke.storage.Storage;
+import duke.storage.TaskList;
+
+public class FunctionStub extends Functions {
 
     public FunctionStub(TaskList ls, Storage st) {
         super(ls, st);
