+++ src/test/java/duke/StorageStub.java
@@ -1,8 +1,10 @@
 package duke;
 
+import duke.storage.Storage;
+
 import java.io.IOException;
 
-public class StorageStub extends Storage{
+public class StorageStub extends Storage {
     public StorageStub(String fp) throws IOException {
         super(fp);
     }
