+++ src/test/java/duke/StorageTest.java
@@ -1,5 +1,6 @@
 package duke;
 
+import duke.storage.Storage;
 import org.junit.jupiter.api.Test;
 import java.io.File;
 import java.io.IOException;
