+++ src/test/java/duke/TaskStub.java
@@ -1,6 +1,8 @@
 package duke;
 
-public class TaskStub extends Task{
+import duke.task.Task;
+
+public class TaskStub extends Task {
 
     public TaskStub(boolean status, String des) {
         super(status, des);
