+++ src/test/java/duke/TaskTest.java
@@ -1,5 +1,6 @@
 package duke;
 
+import duke.task.Task;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
