+++ src/test/java/duke/UiStub.java
@@ -1,6 +1,9 @@
 package duke;
 
-public class UiStub extends UI{
+import duke.functions.Functions;
+import duke.gui.UI;
+
+public class UiStub extends UI {
     public UiStub(Functions fn) {
         super(fn);
     }
